test(ui): skip Qt main window tests on non-Windows; split shell bootstrap tests
- Module-level pytest.skip before PySide6 on Linux CI
- Move build_shell-only tests to test_shell_bootstrap.py for cross-platform coverage

Made-with: Cursor
